chore: addressed review comments
